+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - BUG FIX: Fixed assert when wr8..wr64 had too man args
 - BUG FIX: Fixed bad bool conversion assert
 - BUG FIX: Fixed unlimited const evaluation recursion
+- BUG FIX: Fixed incorrect const substitution
 - INTERNAL: Upgrade to latest md-5 crate.
 - INTERNAL: Upgrade to latest parse_int crate.
 - INTERNAL: Remove unused codespan-reporting crate
@@ -19,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - INTERNAL: Upgrade to latest serial_test crate
 - INTERNAL: Renamed and improved the fuzz_help.md
 - INTERNAL: Added DepthGaurd library for recursion limits
+- INTERNAL: Better linear operand enumeration
 
 ## [6.0.0] - 2026-04-16